Flush after every log line
stdout is line-buffered, so printing a \n will flush the output, but
very often I simply forget it and I wonder why my print statements are
not showing up, thinking my code is broken.
More importantly though when the output it redirected stdout might not
be line-buffered, so that when you are logging into a file for example,
you will only see a log after the program has exited (which is a problem
with the server logs). You might also lose it all, when the program
crashes (and in that case you really want it). So now I just flush every
time.